requirements.txt
transformers>=4.38.0
torch>=2.0.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
numpy>=1.24.0
pandas>=2.0.0
networkx>=3.0
scikit-learn>=1.2.0
streamlit>=1.32.0
google-generativeai>=0.5.0
PyMuPDF>=1.23.0
datasets>=2.18.0
peft>=0.9.0
accelerate>=0.28.0
tqdm>=4.66.0
matplotlib>=3.7.0
pillow>=10.0.0
